requirements.txt
google-adk>=0.5.0
google-generativeai>=0.3.0
litellm>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.0.0